Support more constraints for Conway witnesses
